Refactor/#49 : 폴더 구조 변경
