+++ server/services/core/src/common/exception/chatroom-not-found.exception.ts
