+++ server/services/core/src/common/exception/index.ts
