+++ server/services/core/src/common/type/custom.response.d.ts
