+++ server/services/core/src/common/type/index.ts
