+++ server/services/core/src/domain/chat/controller/chat.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, UseInterceptors } from '@nestjs/common';
 import { MessagePattern, Payload } from '@nestjs/microservices';
-import { LoggingInterceptor } from '../../../interceptor/index';
+import { LoggingInterceptor } from '../../../common/interceptor/index';
 import { ChatService } from '../service/chat.service';
 import { CreateChatDto } from '../dto/create-chat.dto';
 import { UpdateChatDto } from '../dto/update-chat.dto';
