+++ server/services/core/src/provider/prisma/prisma.module.ts
